@@ -2,7 +2,6 @@
   "name": "tanx",
   "version": "0.0.0",
   "license": "MIT",
-  "angular-cli": {},
   "scripts": {
     "ng": "ng",
     "build": "webpack --config server/webpack.config.js && ng build",
@@ -15,48 +14,50 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/common": "^2.4.0",
-    "@angular/compiler": "^2.4.0",
-    "@angular/core": "^2.4.0",
-    "@angular/forms": "^2.4.0",
-    "@angular/http": "^2.4.0",
-    "@angular/platform-browser": "^2.4.0",
-    "@angular/platform-browser-dynamic": "^2.4.0",
-    "@angular/router": "^3.4.0",
-    "@types/ws": "0.0.41",
-    "@uirouter/angular": "^1.0.0-beta.5",
+    "@angular/animations": "^4.2.4",
+    "@angular/common": "^4.2.4",
+    "@angular/compiler": "^4.2.4",
+    "@angular/core": "^4.2.4",
+    "@angular/forms": "^4.2.4",
+    "@angular/http": "^4.2.4",
+    "@angular/platform-browser": "^4.2.4",
+    "@angular/platform-browser-dynamic": "^4.2.4",
+    "@angular/router": "^4.2.4",
+    "@types/ws": "3.0.0",
+    "@uirouter/angular": "^1.0.0-beta.7",
     "bcrypt-nodejs": "0.0.3",
     "body-parser": "^1.17.2",
     "core-js": "^2.4.1",
     "express": "^4.15.3",
     "injection-js": "^2.1.0",
-    "mongodb": "^2.2.26",
-    "mongoose": "^4.10.0",
+    "mongodb": "^2.2.29",
+    "mongoose": "^4.11.0",
+    "ng2-nouislider": "^1.6.1",
+    "nouislider": "^10.0.0",
     "q": "^1.5.0",
-    "rxjs": "^5.0.1",
-    "ts-helpers": "^1.1.1",
-    "ui-router-ng2": "^1.0.0-beta.5",
+    "rxjs": "^5.4.1",
     "ws": "^3.0.0",
-    "zone.js": "^0.7.2"
+    "zone.js": "^0.8.12"
   },
   "devDependencies": {
-    "@angular/cli": "1.0.0-beta.31",
-    "@angular/compiler-cli": "^2.4.0",
-    "@types/jasmine": "2.5.38",
-    "@types/node": "^6.0.42",
-    "codelyzer": "~2.0.0-beta.1",
-    "jasmine-core": "2.5.2",
-    "jasmine-spec-reporter": "2.5.0",
-    "karma": "1.2.0",
-    "karma-chrome-launcher": "^2.0.0",
-    "karma-cli": "^1.0.1",
-    "karma-coverage-istanbul-reporter": "^0.2.0",
-    "karma-jasmine": "^1.0.2",
-    "morgan": "^1.8.1",
-    "protractor": "~5.1.0",
-    "ts-loader": "^2.0.3",
-    "ts-node": "1.2.1",
-    "tslint": "^4.3.0",
-    "typescript": "~2.0.0"
+    "@angular/cli": "1.1.3",
+    "@angular/compiler-cli": "^4.2.4",
+    "@angular/language-service": "^4.2.4",
+    "@types/jasmine": "2.5.53",
+    "@types/node": "~6.0.60",
+    "codelyzer": "~3.1.1",
+    "jasmine-core": "2.6.4",
+    "jasmine-spec-reporter": "~4.1.1",
+    "karma": "~1.7.0",
+    "karma-chrome-launcher": "^2.2.0",
+    "karma-cli": "~1.0.1",
+    "karma-coverage-istanbul-reporter": "^1.3.0",
+    "karma-jasmine": "~1.1.0",
+    "karma-jasmine-html-reporter": "^0.2.2",
+    "protractor": "~5.1.2",
+    "ts-loader": "^2.2.0",
+    "ts-node": "~3.1.0",
+    "tslint": "~5.4.3",
+    "typescript": "~2.3.4"
   }
 }
