@@ -1,30 +1,52 @@
-import { ReplaySubject, Observable } from 'rxjs/Rx';
+import { Subject, Observable } from 'rxjs/Rx';
 import { Injectable } from '@angular/core';
+import { PlayerUnit } from '../objects/classes/player-unit.class';
 import { BaseObject } from '../objects/classes/base-object.class';
+import { User } from '../../auth-forms/user.class';
+import { AuthService } from '../../auth-forms/auth.service';
 
 import { WEBSOCKET_ADDRESS } from '../../../../../config/config';
-import { AuthService } from '../../auth-forms/auth.service';
 
 @Injectable()
 export class UserConnectionService {
     private socket: WebSocket;
+    private subject$ = new Subject();
 
-    private objectsSubject = new ReplaySubject(1);
-    private messagesSubject = new ReplaySubject(1);
-
-    constructor(private AuthService: AuthService) {}
-
-    connect(): void {
-        this.socket = new WebSocket(WEBSOCKET_ADDRESS);
-        this.socket.onopen = () => this.sendUser();
-        this.socket.onmessage = event => this.onMessage(event);
+    constructor(private AuthService: AuthService) {
+        this.connect();
     }
 
-    getObjectsStream(): Observable<BaseObject[]> {
-        return this.objectsSubject.asObservable();
+    getStream(): Observable<[string, any]> {
+        return this.subject$.asObservable();
     }
 
-    sendMessage(method, data): void {
+    getFilteredStream(filter: 'unitUpdates'): Observable<PlayerUnit>;
+    getFilteredStream(filter: 'objectUpdates'): Observable<BaseObject[]>;
+    getFilteredStream(filter: string): Observable<any>;
+    getFilteredStream(filter) {
+        return this.getStream()
+            .filter(([method]) => method === filter)
+            .map(([method, data]) => data);
+    }
+
+    pollFilteredStream(filter: string, interval = 1000, request?: Function, ): Observable<any> {
+        const requestUnitInfo = new Observable(observer => {
+            if (request) { request(); }
+            observer.complete();
+        });
+
+        return this.getFilteredStream(filter)
+            .merge(
+                Observable
+                    .timer(0, interval)
+                    .flatMap(() => requestUnitInfo)
+            );
+    }
+
+    sendMessage(method: 'user', data: User): void;
+    sendMessage(method: 'unitAction', data: { action: string, value: any }): void;
+    sendMessage(method: 'unitUpdates', data?: any): void;
+    sendMessage(method, data?): void {
         const message = { method, data };
         if (this.socket.readyState === WebSocket.OPEN) {
             this.socket.send(JSON.stringify(message));
@@ -39,30 +61,15 @@ export class UserConnectionService {
         this.sendMessage('unitAction', { action, value: false });
     }
 
+    private connect(): void {
+        this.socket = new WebSocket(WEBSOCKET_ADDRESS);
+        this.socket.onopen = () => this.sendUser();
+        this.socket.onmessage = event => this.onMessage(event);
+    }
+
     private onMessage(event): void {
         const { method, data } = JSON.parse(event.data);
-        switch (method) {
-            case 'objectUpdates':
-                this.objectUpdates(data as BaseObject);
-                break;
-            case 'messageUpdates':
-                this.objectUpdates(data);
-                break;
-            case 'roundStart':
-                console.log('Rounds are working on back end: roundStart', data);
-                break;
-            case 'roundStatistics':
-                console.log('Rounds are working on back end: roundStatistics', data);
-                break;
-        }
-    }
-
-    private messageUpdates(data): void {
-        this.messagesSubject.next(data);
-    }
-
-    private objectUpdates(data: BaseObject): void {
-        this.objectsSubject.next(data);
+        this.subject$.next([method, data]);
     }
 
     private sendUser(): void {
