@@ -1,15 +1,15 @@
-export class GameActionsService {
-    private actions: { [name: string]: (sessionId?: string, data?: any) => void } = {};
+import { EventEmitter } from 'events';
 
-    registerAction(name, callback: (sessionId?: string, data?: any) => void): void {
-        this.actions[name] = callback;
-    }
+interface IGameActionsService {
+    on(event: 'user', listener: (sessionId: string, id: string) => void): this;
+    on(event: 'unitAction', listener: (sessionId: string, data: { action: number, value: any }) => void): this;
+    on(event: 'unitUpdates', listener: (sessionId: string, data?: any) => void): this;
 
-    executeAction(name: string, sessionId?: string, data?: any): void {
-        if (this.actions[name]) {
-            this.actions[name](sessionId, data);
-        }
-    }
+    emit(event: 'user', sessionId: string, id: string): boolean;
+    emit(event: 'unitAction', sessionId: string, data: { action: number, value: any }): boolean;
+    emit(event: 'unitUpdates', sessionId: string, data?: any): boolean;
 }
+
+export class GameActionsService extends EventEmitter implements IGameActionsService {}
 
 export default new GameActionsService();
