@@ -14,12 +14,17 @@ export class GameSessionsService {
         this.sessions[sessionId] = socket;
         socket.on('message', message => this.onMessage(sessionId, message));
         socket.on('close', () => this.deleteSession(sessionId));
-        this.GameEventsService.exec('createSession', sessionId);
+        this.GameEventsService.emit('createSession', sessionId);
     }
 
     deleteSession(sessionId: string): void {
         delete this.sessions[sessionId];
-        this.GameEventsService.exec('deleteSession', sessionId);
+        this.GameEventsService.emit('deleteSession', sessionId);
+    }
+
+    sendMessage(sessionId: string, method: string, data?: any): void {
+        const message = JSON.stringify({ method, data });
+        this.postMessage(this.sessions[sessionId], message);
     }
 
     sendAllMessage(method: string, data?: any): void {
@@ -29,9 +34,9 @@ export class GameSessionsService {
         });
     }
 
-    private onMessage(sessionId: string, message: string): void {
-        const { method, data } = JSON.parse(message);
-        this.GameActionsService.executeAction(method, sessionId, data);
+    private onMessage(sessionId: string, message: WebSocket.Data): void {
+        const { method, data } = JSON.parse(message as string);
+        this.GameActionsService.emit(method, sessionId, data);
     }
 
     private postMessage(session: WebSocket, message: string) {
